// src/enum/type.enum.ts
export enum Type {
  INGREDIENTS = 'INGREDIENTS',
  DISH = 'DISH',
  MEAL = 'MEAL',
  DAY = 'DAY',
  USER_GOALS = 'USER_GOALS',
}
export enum EXERCISE {
  LITTLE_OR_NO_EXERCISE = 'LITTLE_OR_NO_EXERCISE',
  LIGHT_ACTIVITY = 'LIGHT_ACTIVITY',
  MODERATE_ACTIVITY = 'MODERATE_ACTIVITY',
  VERY_ACTIVE = 'VERY_ACTIVE',
  EXTRA_ACTIVE = 'EXTRA_ACTIVE',
}
export enum CONDITION {
  diabetes = 'diabetes',
  TIM_MACH = 'TIM_MACH',
  SUY_THAN = 'SUY_THAN',
  BENH_GAN = 'BENH_GAN',
  UNG_THU = 'UNG_THU',
  LOANG_XUONG = 'LOANG_XUONG',
  CELIAC = 'CELIAC',
  GOUT = 'GOUT',
  ANEMIA = 'ANEMIA',
  MALNUTRITION = 'MALNUTRITION',
  RHEUMATOID_ARTHRITIS = 'RHEUMATOID_ARTHRITIS',
  INFLAMMATORY_BOWEL_DISEASE_IBD = 'INFLAMMATORY_BOWEL_DISEASE_IBD',
  COPD = 'COPD',
  PARKINSON = 'PARKINSON',
  ALZHEIMER = 'ALZHEIMER',
  PEPTIC_ULCER_DISEASE = 'PEPTIC_ULCER_DISEASE',
  HEPATITIS_B = 'HEPATITIS_B',
  HEPATITIS_C = 'HEPATITIS_C',
  CIRRHOSIS = 'CIRRHOSIS',
  HYPERTHYROIDISM = 'HYPERTHYROIDISM',
  MULTIPLE_SCLEROSIS_MS = 'MULTIPLE_SCLEROSIS_MS',
  PANCREATITIS = 'PANCREATITIS',
  OSTEOPOROSIS = 'OSTEOPOROSIS',
}